update search page, update advance search reference and remove basic search test
