+++ deployment/frontend/src/pages/dashboard/index.tsx
@@ -33,7 +33,7 @@ export default function index(
             </Layout>
             <Footer
                 links={{
-                    primary: { title: 'Advanced Search', href: '/search' },
+                    primary: { title: 'Search', href: '/search' },
                     secondary: { title: 'Explore Topics', href: '/topics' },
                 }}
                 style="mt-0"
